Passando dados e submissão do pedido
