+++ src/containers/BurguerBuilder/BurguerBuilder.js
@@ -85,35 +85,12 @@ class BurguerBuilder extends Component {
   }
 
   purchaseContinueHandler = () => {
-    {/*this.setState({ loading: true });
-
-    const order = {
-      ingredients: this.state.ingredients,
-      price: this.state.totalPrice,
-      customer: {
-        name: 'John',
-        address: {
-          street: '25 de Dezembro',
-          zipCode: '635272',
-          country: 'Japan'
-        },
-        email: 'test@test.com'
-      },
-      deliveryMethod: 'fastest'
-    }
-
-    axios.post('/orders.json', order)
-      .then(response => {
-        this.setState({ loading: false, purchasing: false });
-      })
-      .catch(error => {
-        this.setState({ loading: false, purchasing: false });
-      }); */}
       const queryParams = [];
 
       for (let i in this.state.ingredients) {
           queryParams.push(encodeURIComponent(i) + '=' + encodeURIComponent(this.state.ingredients[i]));
       }
+      queryParams.push('price=' + this.state.totalPrice);
       const queryString = queryParams.join('&');
       this.props.history.push({
         pathname: '/checkout',
