+++ src/containers/Checkout/Checkout.js
@@ -5,16 +5,23 @@ import ContactData from './ContactData/ContactData';
 
 class Checkout extends Component {
   state = {
-    ingredients: {}
+    ingredients: null,
+    price: null
   }
 
-  componentDidMount() {
+  componentWillMount() {
     const query = new URLSearchParams(this.props.location.search);
     const ingredients = {};
+    let price = null;
+
     for (let param of query.entries()) {
-      ingredients[param[0]] = +param[1]
+      if (param[0] === 'price') {
+        price = param[1];
+      } else {
+        ingredients[param[0]] = +param[1]
+      }
     }
-    this.setState({ingredients: ingredients});
+    this.setState({ingredients: ingredients, price: price});
   }
 
   checkoutCancelledHandler = () => {
@@ -33,7 +40,8 @@ class Checkout extends Component {
           checkoutCancelled={this.checkoutCancelledHandler}
           checkoutContinued={this.checkoutContinuedHandler}/>
         <Routh path={this.props.match.path + '/contact-data'}
-          component={ContactData} />
+          render={() => (<ContactData ingredients={this.state.ingredients}
+              price={this.state.price}/>)} />
       </div>
     )
   }
